+++ irs-edc-client/src/test/java/org/eclipse/tractusx/irs/edc/client/SubmodelFacadeWiremockTest.java
@@ -66,7 +66,7 @@ import org.springframework.web.client.RestTemplate;
 
 class SubmodelFacadeWiremockTest {
     private final static String connectorEndpoint = "https://connector.endpoint.com";
-    private final static String submodelSuffix = "/shells/12345/submodels/5678/submodel";
+    private final static String submodelDataplanePath = "/api/public/shells/12345/submodels/5678/submodel";
     private final static String assetId = "12345";
     private final EdcConfiguration config = new EdcConfiguration();
     private final EndpointDataReferenceStorage storage = new EndpointDataReferenceStorage(Duration.ofMinutes(1));
@@ -134,14 +134,14 @@ class SubmodelFacadeWiremockTest {
             throws EdcClientException, ExecutionException, InterruptedException {
         // Arrange
         prepareNegotiation();
-        givenThat(get(urlPathEqualTo(submodelSuffix)).willReturn(aResponse().withStatus(200)
-                                                                            .withHeader("Content-Type",
+        givenThat(get(urlPathEqualTo(submodelDataplanePath)).willReturn(aResponse().withStatus(200)
+                                                                               .withHeader("Content-Type",
                                                                                     "application/json;charset=UTF-8")
-                                                                            .withBodyFile(
+                                                                               .withBodyFile(
                                                                                     "singleLevelBomAsBuilt.json")));
 
         // Act
-        final String submodel = edcSubmodelClient.getSubmodelRawPayload(connectorEndpoint, submodelSuffix, assetId)
+        final String submodel = edcSubmodelClient.getSubmodelRawPayload(connectorEndpoint, buildWiremockDataplaneUrl(), assetId)
                                                  .get();
 
         // Assert
@@ -196,7 +196,7 @@ class SubmodelFacadeWiremockTest {
                                                                        .authCode("testcode")
                                                                        .properties(Map.of(NAMESPACE_EDC_CID,
                                                                                contractAgreementId))
-                                                                       .endpoint(buildApiMethodUrl())
+                                                                       .endpoint("http://provider.dataplane/api/public")
                                                                        .build();
         storage.put(contractAgreementId, ref);
     }
@@ -206,14 +206,14 @@ class SubmodelFacadeWiremockTest {
             throws EdcClientException, ExecutionException, InterruptedException {
         // Arrange
         prepareNegotiation();
-        givenThat(get(urlPathEqualTo(submodelSuffix)).willReturn(aResponse().withStatus(200)
-                                                                            .withHeader("Content-Type",
+        givenThat(get(urlPathEqualTo(submodelDataplanePath)).willReturn(aResponse().withStatus(200)
+                                                                               .withHeader("Content-Type",
                                                                                     "application/json;charset=UTF-8")
-                                                                            .withBodyFile(
+                                                                               .withBodyFile(
                                                                                     "materialForRecycling.json")));
 
         // Act
-        final String submodel = edcSubmodelClient.getSubmodelRawPayload(connectorEndpoint, submodelSuffix, assetId)
+        final String submodel = edcSubmodelClient.getSubmodelRawPayload(connectorEndpoint, buildWiremockDataplaneUrl(), assetId)
                                                  .get();
 
         // Assert
@@ -225,13 +225,13 @@ class SubmodelFacadeWiremockTest {
             throws EdcClientException, ExecutionException, InterruptedException {
         // Arrange
         prepareNegotiation();
-        givenThat(get(urlPathEqualTo(submodelSuffix)).willReturn(aResponse().withStatus(200)
-                                                                            .withHeader("Content-Type",
+        givenThat(get(urlPathEqualTo(submodelDataplanePath)).willReturn(aResponse().withStatus(200)
+                                                                               .withHeader("Content-Type",
                                                                                     "application/json;charset=UTF-8")
-                                                                            .withBody("test")));
+                                                                               .withBody("test")));
 
         // Act
-        final String submodel = edcSubmodelClient.getSubmodelRawPayload(connectorEndpoint, submodelSuffix, assetId)
+        final String submodel = edcSubmodelClient.getSubmodelRawPayload(connectorEndpoint, buildWiremockDataplaneUrl(), assetId)
                                                  .get();
 
         // Assert
@@ -242,14 +242,14 @@ class SubmodelFacadeWiremockTest {
     void shouldThrowExceptionWhenResponse_400() {
         // Arrange
         prepareNegotiation();
-        givenThat(get(urlPathEqualTo(submodelSuffix)).willReturn(aResponse().withStatus(400)
-                                                                            .withHeader("Content-Type",
+        givenThat(get(urlPathEqualTo(submodelDataplanePath)).willReturn(aResponse().withStatus(400)
+                                                                               .withHeader("Content-Type",
                                                                                     "application/json;charset=UTF-8")
-                                                                            .withBody("{ error: '400'}")));
+                                                                               .withBody("{ error: '400'}")));
 
         // Act
         final ThrowableAssert.ThrowingCallable throwingCallable = () -> edcSubmodelClient.getSubmodelRawPayload(
-                connectorEndpoint, submodelSuffix, assetId).get(5, TimeUnit.SECONDS);
+                connectorEndpoint, buildWiremockDataplaneUrl(), assetId).get(5, TimeUnit.SECONDS);
 
         // Assert
         assertThatExceptionOfType(ExecutionException.class).isThrownBy(throwingCallable)
@@ -260,20 +260,23 @@ class SubmodelFacadeWiremockTest {
     void shouldThrowExceptionWhenResponse_500() {
         // Arrange
         prepareNegotiation();
-        givenThat(get(urlPathEqualTo(submodelSuffix)).willReturn(aResponse().withStatus(500)
-                                                                            .withHeader("Content-Type",
+        givenThat(get(urlPathEqualTo(submodelDataplanePath)).willReturn(aResponse().withStatus(500)
+                                                                               .withHeader("Content-Type",
                                                                                     "application/json;charset=UTF-8")
-                                                                            .withBody("{ error: '500'}")));
+                                                                               .withBody("{ error: '500'}")));
 
         // Act
         final ThrowableAssert.ThrowingCallable throwingCallable = () -> edcSubmodelClient.getSubmodelRawPayload(
-                connectorEndpoint, submodelSuffix, assetId).get(5, TimeUnit.SECONDS);
+                connectorEndpoint, buildWiremockDataplaneUrl(), assetId).get(5, TimeUnit.SECONDS);
 
         // Assert
         assertThatExceptionOfType(ExecutionException.class).isThrownBy(throwingCallable)
                                                            .withCauseInstanceOf(RestClientException.class);
     }
 
+    private String buildWiremockDataplaneUrl() {
+        return buildApiMethodUrl() + submodelDataplanePath;
+    }
     private String buildApiMethodUrl() {
         return String.format("http://localhost:%d", this.wireMockServer.port());
     }