+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- IRS now calls the entire dataplane URL retrieved from the registry href instead of building it from the URL of the EDC token and the path
 
 ## [3.3.2] - 2023-07-31
 ### Fixed