+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - IRS now calls the entire dataplane URL retrieved from the registry href instead of building it from the URL of the EDC token and the path
 
+### Fixed
+- Switched to POST for DTR lookup request
+- Added Base64 encoding to identifier for DTR shell-descriptor request 
+
 ## [3.3.2] - 2023-07-31
 ### Fixed
 - BPN is now passed on correctly when traversing the item graph